Split virtual network APIs to libvirtApi/network.js
